feat(users): move user attributes to a new file
