+++ src/externals/bridge/bridge.service.ts
@@ -3,7 +3,8 @@ import { sign } from 'jsonwebtoken';
 import { Inject, Injectable } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
 import { FileAttributes } from '../../modules/file/file.domain';
-import { User, UserAttributes } from '../../modules/user/user.domain';
+import { User } from '../../modules/user/user.domain';
+import { UserAttributes } from '../../modules/user/user.attributes';
 import { CryptoService } from '../crypto/crypto.service';
 import { HttpClient } from '../http/http.service';
