+++ src/externals/notifications/events/referral-redeemed.event.ts
@@ -1,7 +1,9 @@
 import {
   ReferralAttributes,
-  UserAttributes,
 } from '../../../modules/user/user.domain';
+import {
+  UserAttributes
+} from '../../../modules/user/user.attributes';
 import { Event } from './event';
 
 export class ReferralRedeemedEvent extends Event {
