+++ src/modules/file/file.usecase.ts
@@ -13,7 +13,8 @@ import { BridgeService } from '../../externals/bridge/bridge.service';
 import { FolderAttributes } from '../folder/folder.domain';
 import { Share } from '../share/share.domain';
 import { ShareUseCases } from '../share/share.usecase';
-import { User, UserAttributes } from '../user/user.domain';
+import { User } from '../user/user.domain';
+import { UserAttributes } from '../user/user.attributes';
 import { File, FileAttributes, FileOptions } from './file.domain';
 import { SequelizeFileRepository } from './file.repository';
